+++ packages/bridge-controller/src/bridge-controller.ts
@@ -39,7 +39,6 @@ import {
   getDefaultBridgeControllerState,
   isCrossChain,
   isEthUsdt,
-  isNativeAddress,
   isNonEvmChainId,
   isSolanaChainId,
 } from './utils/bridge.js';